+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 # Release Notes
 
 ## v0.10.0
+* Add a console prefab to be able to log error messages
 * Display selected selected camera or whether stereo mode is active.
 * Delete obsolete files.